Model moved from libs to irohad Update model file
