+++ irohad/CMakeLists.txt
@@ -21,3 +21,4 @@ add_subdirectory(peer_service)
 add_subdirectory(validation)
 add_subdirectory(torii)
 add_subdirectory(network)
+#add_subdirectory(model)
