+++ irohad/model/command.hpp
@@ -14,33 +14,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifndef IROHA_REMOVE_SIGNATORY_HPP
-#define IROHA_REMOVE_SIGNATORY_HPP
 
-#include <common/types.hpp>
-#include <model/command.hpp>
-#include <string>
+#ifndef IROHA_COMMAND_HPP
+#define IROHA_COMMAND_HPP
+#include <ametsuchi/wsv_command.hpp>
+#include <ametsuchi/wsv_query.hpp>
+#include <model/account.hpp>
 
 namespace iroha {
   namespace model {
-
     /**
-     * Attach signatory for account
-     */
-    struct RemoveSignatory : public Command {
-      /**
-       * Destination account to remove from
-       */
-      std::string dst_account;
+      * Abstract Command Model
+      */
+    struct Command {
+      virtual ~Command() = default;
 
-      /**
-       * Public key of signatory to remove.
-       * Note: This public key must be attach to account.
-       * There must be at least two signatories to perform this operation.
-       */
-      ed25519::pubkey_t pubkey;
+      virtual bool validate(ametsuchi::WsvQuery& queries,
+                            const Account& creator) = 0;
+      virtual bool execute(ametsuchi::WsvQuery& queries,
+                           ametsuchi::WsvCommand& commands) = 0;
     };
-  }  // namespace model
-}  // namespace iroha
+  }
+}
 
-#endif  // IROHA_REMOVE_SIGNATORY_HPP
+#endif  // IROHA_COMMAND_HPP
