+++ irohad/model/commands/add_signatory.hpp
@@ -1,0 +1,49 @@
+/**
+ * Copyright Soramitsu Co., Ltd. 2017 All Rights Reserved.
+ * http://soramitsu.co.jp
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef IROHA_ADD_SIGNATURE_HPP
+#define IROHA_ADD_SIGNATURE_HPP
+
+#include <common/types.hpp>
+#include <model/command.hpp>
+#include <string>
+
+namespace iroha {
+  namespace model {
+
+    /**
+     * Attach signatory for account
+     */
+    struct AddSignatory : public Command {
+      /**
+       * Destination account to add new signatory
+       */
+      std::string dst_account;
+
+      /**
+       * New signatory is identified with public key
+       */
+      ed25519::pubkey_t pubkey;
+
+      bool validate(ametsuchi::WsvQuery& queries,
+                    const Account& creator) override;
+      bool execute(ametsuchi::WsvQuery& queries,
+                   ametsuchi::WsvCommand& commands) override;
+    };
+  }  // namespace model
+}  // namespace iroha
+#endif  // IROHA_ADD_SIGNATURE_HPP
