+++ irohad/model/commands/create_asset.hpp
@@ -1,0 +1,42 @@
+/**
+ * Copyright Soramitsu Co., Ltd. 2017 All Rights Reserved.
+ * http://soramitsu.co.jp
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef IROHA_CREATE_ASSET_HPP
+#define IROHA_CREATE_ASSET_HPP
+
+#include <model/model.hpp>
+
+namespace iroha {
+  namespace model {
+
+    /**
+     * Create new asset in the system
+     */
+    struct CreateAsset : public Command {
+      /**
+       * Asset to insert to the system
+       */
+      Asset new_asset;
+
+      bool validate(ametsuchi::WsvQuery& queries,
+                    const Account& creator) override;
+      bool execute(ametsuchi::WsvQuery& queries,
+                   ametsuchi::WsvCommand& commands) override;
+    };
+  }  // namespace model
+}  // namespace iroha
+#endif  // IROHA_CREATE_ASSET_HPP
