+++ irohad/model/model.hpp
@@ -17,49 +17,51 @@
 
 #pragma once
 
+#include <model/account.hpp>
+#include <model/asset.hpp>
 #include <model/block.hpp>
+#include <model/client.hpp>
+#include <model/domain.hpp>
+#include <model/model_crypto_provider.hpp>
+#include <model/model_hash_provider.hpp>
+#include <model/peer.hpp>
 #include <model/proposal.hpp>
 #include <model/transaction.hpp>
 #include <model/transaction_response.hpp>
-#include <model/asset.hpp>
-#include <model/account.hpp>
 #include <model/wallet.hpp>
-#include <model/peer.hpp>
-#include <model/client.hpp>
-#include <model/peer.hpp>
-#include <model/model_crypto_provider.hpp>
-#include <model/model_hash_provider.hpp>
 
 // commands
 #include <model/command.hpp>
+#include <model/commands/add_asset_quantity.hpp>
 #include <model/commands/add_peer.hpp>
 #include <model/commands/add_signatory.hpp>
 #include <model/commands/create_account.hpp>
 #include <model/commands/create_asset.hpp>
-#include <model/commands/create_wallet.hpp>
-#include <model/commands/issue_asset.hpp>
+#include <model/commands/create_domain.hpp>
+#include <model/commands/remove_signatory.hpp>
 #include <model/commands/set_permissions.hpp>
 #include <model/commands/set_quorum.hpp>
 #include <model/commands/transfer_asset.hpp>
 
 // query
-#include <model/query.hpp>
 #include <model/queries/get_account.hpp>
+#include <model/queries/get_account_assets.hpp>
 #include <model/queries/get_signatories.hpp>
 #include <model/queries/get_transactions.hpp>
-#include <model/queries/get_wallets.hpp>
+#include <model/query.hpp>
 
 // query response
-#include <model/query_response.hpp>
 #include <model/queries/responses/account_response.hpp>
 #include <model/queries/responses/error_response.hpp>
 #include <model/queries/responses/signatories_response.hpp>
 #include <model/queries/responses/transactions_response.hpp>
 #include <model/queries/responses/wallets_response.hpp>
+#include <model/query_response.hpp>
 
 
 /**
- * Model module provides objects that are useful for all other modules in system.
+ * Model module provides objects that are useful for all other modules in
+ * system.
  * Model objects do not depend on transport, such as protobuf.
  * Model objects in general are structures with public immutable fileds.
 */
