+++ irohad/model/model_hash_provider_impl.hpp
@@ -13,26 +13,28 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef IROHA_ASSET_HPP
-#define IROHA_ASSET_HPP
 
+#ifndef IROHA_MODEL_HASH_PROVIDER_IMPL_HPP
+#define IROHA_MODEL_HASH_PROVIDER_IMPL_HPP
+
+#include <algorithm>
 #include <common/types.hpp>
-#include <string>
+#include <crypto/base64.hpp>
+#include <crypto/crypto.hpp>
+#include <crypto/hash.hpp>
+#include <model/model_hash_provider.hpp>
 
 namespace iroha {
   namespace model {
+    class HashProviderImpl : public HashProvider<ed25519::pubkey_t::size()> {
+     public:
+      iroha::hash256_t get_hash(const Proposal &proposal) override;
 
-    /**
-     * Asset Data Model
-     */
-    struct Asset {
+      iroha::hash256_t get_hash(const Block &block) override;
 
-      /**
-       * Asset unique identifier
-       */
-      const std::string name;
+      iroha::hash256_t get_hash(const Transaction &tx) override;
     };
   }
 }
 
-#endif  // IROHA_ASSET_HPP
+#endif  // IROHA_MODEL_HASH_PROVIDER_IMPL_HPP
