+++ irohad/model/queries/get_account_assets.hpp
@@ -1,0 +1,35 @@
+/**
+ * Copyright Soramitsu Co., Ltd. 2017 All Rights Reserved.
+ * http://soramitsu.co.jp
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef IROHA_GET_WALLETS_HPP
+#define IROHA_GET_WALLETS_HPP
+
+#include <model/model.hpp>
+#include <string>
+
+namespace iroha {
+  namespace model {
+
+    /**
+     * Query for get all account's wallet
+     */
+    struct GetAccountAssets : Query {
+      std::string account_name;
+    };
+  }  // namespace model
+}  // namespace iroha
+#endif  // IROHA_GET_WALLETS_HPP
