+++ irohad/model/queries/get_wallets.hpp
