+++ irohad/model/queries/responses/error_response.hpp
