+++ irohad/model/queries/responses/wallets_response.hpp
