+++ irohad/model/query.hpp
@@ -1,0 +1,41 @@
+/**
+ * Copyright Soramitsu Co., Ltd. 2017 All Rights Reserved.
+ * http://soramitsu.co.jp
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef IROHA_QUERY_HPP
+#define IROHA_QUERY_HPP
+
+#include <model/model.hpp>
+#include <string>
+
+namespace iroha {
+  namespace model {
+    /**
+     * This model represents user intent for reading ledger.
+     * Concrete queries should extend this interface.
+     */
+    struct Query {
+
+      /**
+       * Signature of query's creator
+       */
+      Signature signature;
+
+      virtual ~Query() {}
+    };
+  } //namespace model
+} //namespace iroha
+#endif //IROHA_QUERY_HPP
